Create a feed for simulated news
This commit creates the FeedModule. It will serve up a new feed item every 7 to 12 ticks. Debug out has been modified to support this new feature, but history is only kept for this output purpose. Clients receive a single feed post when the web socket endpoint is invoked. This news has no affect on the stock market...yet.